The Big Merge pt2

+++ writeUninstallInstructions.sh
@@ -1,12 +1,15 @@
+#!/bin/bash
+: "${bindir:?}"
+: "${POPLOG_HOME_DIR:?}"
 cat << ****
 Instructions to Fully Uninstall Poplog
 ======================================
 
 Poplog is installed in a single folder and it is sufficient to remove that
-folder and the single symlink that points into it. So, from inside a terminal, 
+folder and the single symlink that points into it. So, from inside a terminal,
 type the following commands.
 
-    sudo rm ${bindir}/poplog 
+    sudo rm ${bindir}/poplog
     sudo rm -rf ${POPLOG_HOME_DIR}
 
 And that's it.
